app/requirements.txt
aniso8601==6.0.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
